split server file
